Minor change.
Removed Duplicate Packet Detection in onCharacteristicChanged in attempt
to always send ACK packet to xBridge, even if it is a duplicate.
Sometimes, as this is not sent immediately, the wixel does not get the
ACK packet, and continues to stay awake for 2 minutes.  This should
prevent this.

A pseudo Duplicate Packet Detection, or packet throttling, still occurs in DexCollectionService.