Adds support for a http method of retrieving json data.
This is very similar to the existing json service used by "wifi" wixels
but works with http and https so that data can be received from a cloud
web service instead of a custom server.

The main purpose for this is to eliminate the need to provide hosting
for the parakeet xdrip_parakeet_gsm_receiver.py on a dedicated server
and instead be able to run it within a sandboxed web instance on
a cloud provider.