Restrict data fetched for ongoing Notification
This restricts the data fetched for the ongoing notification to the data that is needed.
It will not fetch the whole day but just the 3 hours shown in the notification.

This also fixes the bug that values hours before the values shown in the chart affect how it is stretched on the y axis.