Allow history retrieval from a Blukon device, when connecting to the device after a period of more than 15 minutes has elapsed, without using the LibreOOPAlgorithm.
(Note that this includes a fix for a possible bug in the LibreOOPAlgorithm code which occurs when the sensorAge is requested before the history. The sensor age is updated, and then the history is handed off to LibreAlarmReceiver, which again checks the sensor age. When finding that the sensor age is current, it discards the historical data.)