Add poll delay for Nightscout Follow
Enabling a poll delay for Nightscout follow to reduce the time from
a new value is made available on Nightscout until xdrip polls for new
data. Depending on the setup of nightscout this might be anything
between 0 and 150 seconds. Minimum value is set to 0s. Max is 300s.

This pull request fixes #929 by allowing custom poll delay.