#570 sensor age alert
Add a configuration option to alert the user when the sensor age reaches a pre-defined value (in days).
Serves as a reminder to plan the sensor replacement, or place a sensor in advance (e.g. Libre).
The alternative would be to set a reminder elsewhere (using another app), but since xDrip must be informed when a sensor is started anyway, this is a convenient and easy option.