Add functions for converting between timestamps and X positions on the graph
This commit introduces functions that replace the division / multiplication by
the constant `FUZZER` with function calls. The replacement functions also shift
the zero time point of the timestamps to one year in the past, instead of the
fixed 1970/1/1. This is important, as the graphing library uses single precision
floats internally, and in 2020, a single-precision float is only capable of
representing a standard UNIX timestamp divided by `FUZZER` with a precision of
approximately 2.5 min. By shifting the values closer to the origin, we gain enough
precision to e.g. display raw Libre2 values at the time that they were actually
received.