Allow xDrip to read libredata using rest instead of mongo.
As of now mlab does not exist anymore, and android apps can not read from mongo directly anymore.
So, I have added a rest api to read libre data.

There is also an example rest service that can be used to read from atlas.
Can be forked from: https://github.com/tzachi-dar/node-js-getting-started.git

The format of the url should be: http://127.0.0.1:5000/pass/<pass>

An example of the data returned is:
[{"_id":"5fd6912c13138ae253e770a1","BlockBytes":"DpSQFwMAAAAAAAAAAAAAAAAAAAAAAAAADAYKG0UFyNyZAEcFyNjZAD4FyKDZAEcFyIDZAEUFyCTZACUFyASZABkFyMzYAA0FyKgYAegEyNDYANkEyPBYADQFyBTaADsFyPjZAD8FyBBaAEgFyBCaAE8FyOxZAEsFyMxZAOkEiG4bgFEFyGhbgKEFyJDZACMGyIBYAbgFyNRaAMkDyPTaAIEDyMAaAZQEyJwaAaUFyEhbAEsFyKxbAAAEyKSaAH8DyBxbAKQDyEgbgI0DyMhaAGkDyMyaAKQDyJCaAKgDyKwbgGADyCwcgOsByDwbgEQCyPwagK0CyJgbgFICyJSaAHICyFxaADgDyDyaABMEyNQZARcFyAhaADQFyCTZALcEyIQbgL0FyJQbgPUEyDgbgFgEyDQbgHwEyHQagHtOAAC3yQAI7gm9UBQHloBaAO2mBo0ayARL2GY=","CaptureDateTime":1607897388600,"ChecksumOk":1,"DebugInfo":"rpi4-1 14-12-2020 00:09:48 tomato","TomatoBatteryLife":47,"UploaderBatteryLife":100,"Uploaded":0,"HwVersion":"1001","FwVersion":"4","SensorId":"0M000A61P94","NoSensor":0}]