Add more logging to help users understand why libre 2 does not communicate with bluetooth.
I have received a few complaints that xDrip does not work with libre2.
Usually this things are solved when the user retries the connection a few times.
This change high level logging weather scanning the sensor and enabling BT has worked correctly, in order to make users know weather they enabled the sensor successfully or not.