Use an improved version of the active android library.
The new library is only changed in a few lines (see https://github.com/tzachi-dar/ThreadSafeActiveAndroid) and contains the option to use indexed by.
This allows us to better choose the index that is used for queries, and with this new index we can execute queries faster (see the next checkin as an example).

I have tried Using ANALYZE on the db and this did not have an effect on the indexes that have been used.
The simplest way to check this is to use SQLite browser and run explain query (for example explain query plan SELECT * from BgReadings where sensor = "xx" and  timestamp <= 1000 and  calculated_value = 0 and raw_calculated = 0 ORDER by timestamp DESC)
Then run analyze on the db and run the query again.
As a side note, the analyze command only creates a small table so it is not possible for it to really capture the situation on the DB.