Don't allow xDrip to work with OOP2 versions before 1.2
There have been two versions before the 1.2.
The first version only supported raw calibration mode. Users of this version will not see any difference, and if they try to move to a new mode will get a message saying that they need to upgrade OOP2, or return to "calibrate based on raw" mode.

The second version had a bug and was removed. Still some users might be using it in "no calibration" or "calibrate based on glucose" mode. This users will see xDrip stops working, but this is the preferred way since the banned OOP2 had a bug that might make it show BG values that are very inaccurate (300 instead of 100).
OOP2 does not have an auto-update feature, so this is the only way to revoke it.