Fix intermittent crashes during value and noise estimation
Both the Savitzky-Golay filter and the polynomial noise estimator
assumed that the measurements are input in chronological order, which
used to be true when I wrote the code. The upstream code changed, and
now the values are sometimes not sorted anymore, which caused a crash in
an assertion that I included to make sure the incorrectly ordered
measurement sets do not get used in the program.

This patch fixes the problem by simply sorting the arrays with the
measurements chronologically before doing the calculations.

This should fix the intermittent application crashes when using the
Savitzky-Golay filter and / or the polynomial noise estimator.