First commit to use LTE.
The long term goal is to allow xdrip to show the BG (for followers) even when no phone is near by. This should be done by following a nightscout web site.

It seems that the only library that allows a wear os watch to do http queries is cronet (see https://developer.android.com/guide/topics/connectivity/cronet). Using cronet requires that one will use 'com.google.android.gms:play-services-cronet:18.0.1' which in turn requires play-services-wearable to be updated to version 18.0.0
This forces the use of AndroidX.
And here we are with this change.
So, right now, it only goes to example.com and gets a page from there, but we have many  changes caused by this.