This change changes the behavior when one is scanning a sensor that is not the current sensor.
The old behavior was to close the old sensor in the case that a new sensor was scanned.
The new behavior is this:
If the scan was a manual scan, we ignore the scan and return (this is the reason that the check has moved to the first possible position).
If this is a reading from a device like miaomiao the old behavior is preserved (close the sensor).
The reason for the difference is that if one is using a device, than he has already sticked the device to the sensor, and there is no going back.